reference/crewai-baseline/requirements.txt
langchain-exa==0.0.1
exa_py==1.0.7
python-dotenv==1.0.1
crewai==0.130.0
requests==2.31.0
